Change encoding of enums.
In order to be able to skip over unknown enums, we need to put them in
a container. During decoding we do not know if an unknown type is a
struct or enum and can make no distinction. So we can always only skip
a single CBOR item, hence every enum variant is now represented as a
2-element array--index and variant data.

Signed-off-by: Toralf Wittner <tw@dtex.org>