Fix lifetime handling when deriving `Decode`.
So far every lifetime has been constrained to the decoding input's
lifetime even when the type did not actually borrow from the input.
The new approach is to implicitly borrow `&str`, `&[u8]` and `Option`s
of those. All other lifetimes are unconstrained and borrowing needs to
be explicitly enabled using `#[b(...)]` instead of `#[n(...)]` as the
index attribute. For `Cow` in particular `#[b(...)]` switches from its
`Decode` impl which produces owned values to decoding of the inner type
which is then wrapped in `Cow::Borrowed`.

In addition uninhabited enums are now handled correctly unit structs are
correctly decoded.

Finally the documentation now explains the resulting CBOR format and the
details regarding borrowing.

Signed-off-by: Toralf Wittner <tw@dtex.org>