Allow customising the encoding.
Up to now, structs and variants have always been encoded as CBOR maps
which is an efficient encoding for sparse types (types with many
`None`s). It does however incur an overhead of at least one byte per
field as it needs to encode the index. Dense types pay this overhead
too even though they do not benefit from it. This includes in
particular small and stable types.

This commit therefore changes the default encoding from maps to arrays.
Under array encoding indexes are not part of the encoding but derived
from the array position of a field value. It does mean that `None`s
end up as `null` in the encoding, but as long as their number is small
the price may still be lower than with the overhead of explicit index
encoding.

Nevertheless for types which are sparse by design or expected to
undergo many changes, the map encoding would still be the better option
so this commit also adds an attribute to select which encoding to
choose per struct, enum and enum variant. Possible values are
`#[cbor(array)]` for the array based encoding (which is used by default)
and `#[cbor(map)]` for selecting the map encoding.

Signed-off-by: Toralf Wittner <tw@dtex.org>