Add `bytes` module.
To supports optimised encoding of `&[u8]` and `Vec<u8>`.
Removes the `Decode` impl of `&[u8]` as it is not matching the
`Encode` implementation which is subsumed by `[T]` and produces
CBOR arrays. In the absence of trait impl specialisation it
seems best to delegate bytes handling to specialised types.

For context see #4.

Signed-off-by: Toralf Wittner <tw@dtex.org>