Add `index_only` attribute.
This supports a more compact encoding for enums which do not have any
fields and are not expected to ever have some. By only encoding the
variant index the encoding of those enums can be much more compact,
using at best only 1 byte instead of 3. Of course, this encoding is
not compatible with the regular enum encoding, so it should be used
with caution.

Signed-off-by: Toralf Wittner <tw@dtex.org>