Add optional support for skipping indefinite arrays and maps.
As pointed out by @Hawk777 in issue #9, the current implementation of
`Decoder::skip` does not properly support indefinite maps or arrays.
Furthermore, correctly supporting these requires recursion or an
internal control stack. This commit adds a feature `"skip-any"` which
implements `Decoder::skip` using an allocating control stack, similarly
to `minicbor::display`. The feature is opt-in because indefinite arrays
or maps are often not needed and the implementation is less performant
than the one not supporting those.

Signed-off-by: Toralf Wittner <tw@dtex.org>