Replace `std::borrow::Cow` with `minicbor::bytes::Cow`.
`minicbor-derive` may be used in a `no_std` setting with just `alloc`
enabled, hence we can not always refer to `std::borrow::Cow`. OTOH `std`
does not re-export `alloc` like it does for `core` so we better use a
stable reference regardless of whether we are in `no_std` or not.

Signed-off-by: Toralf Wittner <tw@dtex.org>