Change encoding of IP addresses.
The current encoding format encodes the address octets as an array of
`u8` values instead of using more compact CBOR bytes. Using the latter,
IPv6 addresses can be almost half the size, hence this commit switches
the format to CBOR bytes.

Because this is a breaking change, affecting

- std::net::IpAddr
- std::net::Ipv4Addr
- std::net::Ipv6Addr
- std::net::SocketAddr
- std::net::SocketAddrV4
- std::net::SocketAddrV6

a new module `minicbor::legacy` is introduced which contains newtype
wrappers for the aforementioned types which continue to use the
current, array-based encoding. Users can opt out of the new compact
format by enabling the cargo feature `legacy` and importing the types
from the legacy module.

Signed-off-by: Toralf Wittner <tw@dtex.org>