Rework `encode::write::Write` impls.
Currently the `Write` trait has a blanket impl for types that implement
`std::io::Write`, which includes `Vec<u8>` and also `&mut W` for any
`W` that implements `std::io::Write`.

In a no_std setting, none of these impls are available and neither is
`std::io::Error`. The impl for `Vec<u8>` that is implemented if feature
"alloc" is given thus has `core::convert::Infallible` as error type.
This discrepancy is unfortunate and in addition functions like `to_vec`
require the "std" feature, even though "alloc" would be enough.

This commit changes the implementation as follows:

1. The blanket impl for types that implement `std::io::Write` is removed.
2. The impl for `Vec<u8>` requires "alloc" and does not change for "std".
3. `minicbor::{to_vec, to_vec_with}` are available with "alloc" and
   return as error type `encode::Error<Infallible>`.
4. An adapter `Writer` is provided that wraps any type that implements
   `std::io::Write` and implements `Write`. This can be used instead of
   the previous blanket impl.
5. As a consequence, `minicbor_io::Error::Encode` now also contains an
   `encode::Error<Infallible>` instead of an `encode::Error<std::io::Error>`.
6. A type `Cursor` is added that implements `Write` for `&mut [u8]`,
   `[u8; N]` and `Box<[u8]>`. It largely mimicks `std::io::Cursor` and
   allows tracking of write positions.

Signed-off-by: Toralf Wittner <tw@dtex.org>