Rework tags.
The current `Tag` enum is a closed set of values and stores all unknown
tag values in `Tag::Unassigned`. However these values may actually
correspond to registered tag values at IANA so the name "unassigned" is
somewhat unfortunate.

To better represent the openess of tags I propose to make `Tag` a mere
newtype of `u64` with conversions to and from the underlying integer
value. In addition, a new type `IanaTag` represets the set of registered
tags. It only has constructors for known values and is marked as
`#[non_exhaustive]` to allow the addition of constructors in the future.

With these two types in place one could either operate on raw tag values
or attempt the conversion to `IanaTag` for a more convenient interface.

Signed-off-by: Toralf Wittner <tw@dtex.org>