Add "borrow" attribute.
Similar to serde's attribute of the same name it can be attached to
fields. The existing `#[b(_)]` attribute is mututally exclusive with
`#[cbor(borrow)]` and can be thought of as syntactic sugar for writing
`#[cbor(n(_), borrow)]`.

Note that `borrow` accepts optional lifetime paramters to restrict the
lifetimes that should be constrained.

Signed-off-by: Toralf Wittner <tw@dtex.org>